requirements.txt
affine==2.4.0
attrs==25.4.0
certifi==2025.11.12
charset-normalizer==3.4.4
click==8.3.1
click-plugins==1.1.1.2
cligj==0.7.2
contourpy==1.3.3
cycler==0.12.1
fonttools==4.60.1
idna==3.11
kiwisolver==1.4.9
matplotlib==3.10.7
numpy==2.3.5
packaging==25.0
pillow==12.0.0
pyparsing==3.2.5
python-dateutil==2.9.0.post0
PyYAML==6.0.3
rasterio==1.4.3
requests==2.32.5
six==1.17.0
urllib3==2.5.0
